import SignRequest after moving to another package
